+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Intercom Integration - Prefer Intercom accessToken configuration to old fashioned appId/apiKey.
+- Intercom Integration - Remove support for old configuration parameter use "userCollection" (use mapping instead).
 
 ## RELEASE 1.1.7 - 2017-06-13
 ### Changed